Catch a more general exception when checking for rotation from content resolver
The behaviour of a content provider when it is queried for a column that does
not exist is not well defined. Content providers can throw any one of the
exceptions defined in Parcel as well as a number SQLiteExceptions via
DatabaseUtils.readException()/.writeException(). The native Android content
provider will throw an SQLiteException if the orientation column is not found.