reassign the PicassoDrawable to the ImageView after changing it
When the Drawable is set, the ImageView caches the Drawable's height
and width and precomputes the scaling matrix.  PicassoDrawable can
change height and width (if the placeholder, error, or bitmap have
different sizes) after being assigned to the ImageView, leading to
incorrect scaling.

This commit changes the PicassoDrawable static setBitmap() and
setPlaceHolder() methods to reassign the drawable to the target
ImageView, forcing the scaling matrix to be properly computed.