change PicassoDrawable to respect the bounds given by its container
PicassoDrawable scales the bounds it passes to the underlying
BitmapDrawable. Thus, when the bitmap is drawn, it does not respect
the bounds given by the parent (e.g., ImageView), but instead the
scaled bounds.  This breaks features like the ImageView scaleType
attribute.

This commit changes PicassoDrawable to always respect the bounds given
to it (e.g., by the parent ImageView).  Scaling should be applied in
the standard way---by calling setBounds()---and the given bounds are
passed directly to the underlying BitmapDrawable. Thus,
PicassoDrawable should always render the same as a pure
BitmapDrawable.