Fix potential Message leak
Introduces a local variable for the Message in CleanupThread, and makes sure to clean it after sending the message.

Here's why:

* I have a heapdump that I took on Genymotion with Android 4.4.4.
* It shows that CleanupThread is keeping a reference to a Message object, long after it's been sent to the handler (and therefore recycled).
* Picasso.shutdown is false (and there's only one Picasso instance). There's also only one CleanupThread instance.
* The reference queue is empty, which means that CleanupThread is blocking on `referenceQueue.remove();`
* So I'm assuming that somehow the message got pushed in a local variable, and won't be cleared until `remove()` unblocks.

This has caused a memory leak when some other code is obtaining a message and then holding on to it (not recycling it). If such code sets `obj` on the message, `obj` will not be GCed until the next Picasso request is GCed.

Call me crazy... I have a heapdump though :) .