Replace MarkableInputStream with a superior Okio version.
This new InputStream uses the Buffer of the supplied BufferedSource to ensure it has buffered the necessary bytes to fulfill the reads of the consumer (via copies). This allows the original source to still be consumed normally by virtue of the fact that we never read directly from it.

There's a host of tests that we need for this but I'm holding off on writing them because we might actually upstream a mechanism similar to this into Okio.