Remove batching to the main thread.
Texture uploads to the GPU are a fixed cost and by batching we actually group that cost which might actually cause missed frame deadlines. Multiple posts in the same frame isn't really an issue either as layout is already batched. Also this was speculatively added without empirical data to justify its presence.